refactor: updateStateGarage
