+++ async-race/src/index.ts
@@ -1,4 +1,4 @@
-import { UpdateGarage } from './components/ui/updateGarage';
+import { UpdateStateGarage } from './components/ui/updateStateGarage';
 import { Listener } from './components/ui/listener';
 import { App } from './components/app/app';
 import './styles.scss';
@@ -8,5 +8,5 @@ const root = document.querySelector('#root');
 if (!root) throw new Error(`root = ${typeof root}`);
 
 new App(root).render();
-await new UpdateGarage().render();
+await new UpdateStateGarage().render();
 new Listener().start();
